pyproject.toml
[project]
name = "copilot-chan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-adk>=0.3.1",
    "fastapi==0.118.3",
    "firebase-admin>=6.9.0",
    "google-adk>=1.16.0",
    "ruff>=0.14.1",
]
